Move ix-json into qntx-plugins/ for dev server hot-reload
The dev server transpiles plugin .ts modules on the fly from
qntx-plugins/{name}/web/. ix-json lived outside that convention
at the repo root, so edits required a full plugin rebuild.

Also fixes double-wrapping in glyph-module-loader — the loader
was creating its own canvasPlaced container around the module's
container, resulting in an empty glyph.
