+++ .gitignore
@@ -31,7 +31,7 @@ web/coverage/
 *.wasm
 qntx-code-plugin
 qntx-webscraper-plugin
-qntx-ix-json/qntx-ix-json-plugin
+qntx-plugins/ix-json/qntx-ix-json-plugin
 
 # IDE
 .vscode/
