+++ qntx-plugins/ix-json/tsconfig.json
